move back some local types to web3-eth-contract
