+++ packages/web3-types/src/eth_contract_types.ts
@@ -15,106 +15,10 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { Address, BlockNumberOrTag, HexString32Bytes, Uint } from './eth_types';
-import { AbiFragment, ContractAbi } from './eth_abi_types';
+import { Address, Uint } from './eth_types';
 import { SupportedProviders } from './web3_base_provider';
 import { Bytes, HexString } from './primitives_types';
 import { EthExecutionAPI } from './apis/eth_execution_api';
-
-export type ContractAbiWithSignature = ReadonlyArray<AbiFragment & { signature: HexString }>;
-
-export interface EventLog {
-	readonly event: string;
-	readonly id?: string;
-	readonly logIndex?: bigint | number | string;
-	readonly transactionIndex?: bigint | number | string;
-	readonly transactionHash?: HexString32Bytes;
-	readonly blockHash?: HexString32Bytes;
-	readonly blockNumber?: bigint | number | string;
-	readonly address: string;
-	readonly topics: HexString[];
-	readonly data: HexString;
-	readonly raw?: { data: string; topics: unknown[] };
-	readonly returnValues: Record<string, unknown>;
-	readonly signature?: HexString;
-}
-
-export interface ContractEventOptions {
-	/**
-	 * Let you filter events by indexed parameters, e.g. `{filter: {myNumber: [12,13]}}` means all events where `myNumber` is `12` or `13`.
-	 */
-	filter?: Record<string, unknown>;
-	/**
-	 * The block number (greater than or equal to) from which to get events on. Pre-defined block numbers as `earliest`, `latest`, `pending`, `safe` or `finalized` can also be used. For specific range use {@link Contract.getPastEvents}.
-	 */
-	fromBlock?: BlockNumberOrTag;
-	/**
-	 * This allows to manually set the topics for the event filter. If given the filter property and event signature, (topic[0]) will not be set automatically. Each topic can also be a nested array of topics that behaves as “or” operation between the given nested topics.
-	 */
-	topics?: string[];
-}
-
-export interface ContractOptions {
-	/**
-	 * The maximum gas provided for a transaction (gas limit).
-	 */
-	readonly gas?: Uint;
-	/**
-	 * The gas price in wei to use for transactions.
-	 */
-	readonly gasPrice?: Uint;
-	/**
-	 * The address transactions should be made from.
-	 */
-	readonly from?: Address;
-	/**
-	 * The byte code of the contract. Used when the contract gets {@link Contract.deploy | deployed}
-	 */
-	readonly input?: Bytes;
-	/**
-	 * The {@doclink glossary/json_interface | json interface} object derived from the [ABI](https://github.com/ethereum/wiki/wiki/Ethereum-Contract-ABI) of this contract.
-	 *
-	 * Re-setting this will regenerate the methods and events of the contract instance.
-	 *
-	 * ```ts
-	 * myContract.options.jsonInterface;
-	 * > [{
-	 *     "type":"function",
-	 *     "name":"foo",
-	 *     "inputs": [{"name":"a","type":"uint256"}],
-	 *     "outputs": [{"name":"b","type":"address"}],
-	 * 	   "signature": "0x...",
-	 * },{
-	 *     "type":"event",
-	 *     "name":"Event",
-	 *     "inputs": [{"name":"a","type":"uint256","indexed":true},{"name":"b","type":"bytes32","indexed":false}],
-	 * 	   "signature": "0x...",
-	 * }]
-	 *
-	 * // Set a new ABI interface
-	 * // Note: the "signature" of every function and event's ABI is not needed to be provided when assigning.
-	 * //	It will be calculated and set automatically inside the setter.
-	 * myContract.options.jsonInterface = [...];
-	 * ```
-	 */
-	get jsonInterface(): ContractAbiWithSignature;
-	set jsonInterface(value: ContractAbi);
-
-	/**
-	 * The address used for this contract instance. All transactions generated by web3.js from this contract will contain this address as the `to`.
-	 *
-	 * The address will be stored in lowercase.
-	 *
-	 * ```ts
-	 * myContract.options.address;
-	 * > '0xde0b295669a9fd93d5f28d9ec85e40f4cb697bae'
-	 *
-	 * // set a new address
-	 * myContract.options.address = '0x1234FFDD...';
-	 * ```
-	 */
-	address?: Address; // All transactions generated by web3.js from this contract will contain this address as the "to".
-}
 
 export interface ContractInitOptions {
 	/**
